src/requirements.txt
google-adk
requests
# yfinance==0.2.56
# psutil==5.9.5
# litellm==1.66.3
# google-generativeai==0.8.5
# python-dotenv==1.1.0
kafka-python
Flask
firebase_admin
google-cloud-aiplatform
cloudpickle
langchain
langchain-core
langchain-google-vertexai
pydantic
